build: add tarball mode that disables build
This mode is intended to be used by CI.
When the environment defines a `USE_TARBALL` variable containing a path
to a .tar.gz file generated by `make pack`, a new target `make unpack`
is available and the rules for making dist files are replaced by a
warning. This is useful to run the test targets without re-generate the
dist files.